requirements.txt
paho-mqtt==1.5.0
Pillow==7.1.2
